fix: reorganization slices folders
