+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,7 +1,13 @@
 import { Button, CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-components/dist/index.js";
 import styles from './burger-constructor.module.css';
 import { useDispatch, useSelector } from 'react-redux';
-import { orderIngredients, orderCost, placeAnOrder, orderIngredientsId, deleteIngredient, showOrder, mainIngredients } from '../../services/redux/order-slice';
+import { orderIngredients, 
+         orderCost, 
+         placeAnOrder, 
+         orderIngredientsId, 
+         deleteIngredient, 
+         showOrder, 
+         mainIngredients } from '../../services/redux/order-slice/order-slice';
 import { useDrop } from 'react-dnd';
 import { useHistory } from 'react-router-dom';
 import OrderItem from '../order-item/order-item';
