+++ src/services/redux/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
-import ingredientsSlice from "./ingredients-slice";
-import orderSlice from "./order-slice";
-import authorizationSlice from "./authorization-slice";
+import ingredientsSlice from "./ingredients-slice/ingredients-slice";
+import orderSlice from "./order-slice/order-slice";
+import authorizationSlice from "./authorization-slice/authorization-slice";
 
 const rootReducer = combineReducers({
   ingredientsSlice: ingredientsSlice,
